Avoid computing transitive coming edges
Instead of always computing transitive incoming edges, determine if there are such nodes
as they are added to the edge state. The reason is that the creation of a list turns out
to be a hotspot during dependency resolution. We also cache the value of "transitive"
because the JIT isn't able to inline delegated calls because of megamorphic calls.