Use `LinkedHashSet` instead of `LinkedList`
While in general we avoid use of linked hashset, in this case it proves to be useful because it's
pretty common to remove edges from the list. Sometimes, the list contains numerous elements but
not the one that we remove, making it rather inefficient to use a linked list. Array list would also
be expensive, so we come back to the good old linked hash set.